Fix for the new testing environment
